requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
dnspython==2.8.0
idna==3.13
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-whois==0.9.6
requests==2.33.1
rich==15.0.0
six==1.17.0
urllib3==2.6.3
